Refactor API tests into focused, maintainable files
Separate concerns following Extract + Test + Commit methodology:

Files created:
- test_mocks.go: Shared mocks and test utilities (MockJobManager, MockDBClient)
- jobs_create_test.go: POST /v1/jobs endpoint tests
- dashboard_endpoints_test.go: Dashboard stats/activity tests
- webhook_endpoints_test.go: Webflow webhook integration tests

Benefits:
✅ 889-line monolithic file → 4 focused files (avg ~200 lines each)
✅ Shared mocks eliminate duplication
✅ Clear separation of endpoint concerns
✅ Easier maintenance and extension
✅ All tests maintain 24.5% coverage

Following established testing patterns in codebase (build_task_query_test.go, etc.)
